Improve Endpoint Management UI
