+++ console/workspaces/pages/mcp-proxies/package.json
@@ -39,12 +39,14 @@
     "@agent-management-platform/api-client": "workspace:*",
     "@agent-management-platform/types": "workspace:*",
     "@agent-management-platform/shared-component": "workspace:*",
-    "zod": "4.3.6"
+    "zod": "4.3.6",
+    "lodash": "4.17.21"
   },
   "devDependencies": {
     "@types/node": "24.6.0",
     "@types/react": "19.2.14",
     "@types/react-dom": "19.1.9",
+    "@types/lodash": "4.17.13",
     "@typescript-eslint/eslint-plugin": "8.33.1",
     "@typescript-eslint/parser": "8.33.1",
     "eslint": "9.36.0",
